Simplifies commands
