+++ .github/workflows/changelog-generator.yml
@@ -62,7 +62,7 @@ jobs:
                   DEBUG: ${{ inputs.debug == 'true' && 'changelog*' || '' }}
               run: |
                   if [ "${{ inputs.dry_run }}" = "true" ]; then
-                      pnpm changelog:sync:all --dry-run
+                      pnpm changelog:publish --dry-run
                   else
-                      pnpm changelog:sync:all
+                      pnpm changelog:publish
                   fi
