+++ package.json
@@ -19,7 +19,6 @@
     "write-translations": "docusaurus write-translations",
     "write-heading-ids": "docusaurus write-heading-ids",
     "changelog:entry:new": "tsx packages/changelog-generator/src/index.ts entry-new",
-    "changelog:sync:all": "tsx packages/changelog-generator/src/index.ts sync-all",
     "changelog:compile:json": "tsx packages/changelog-generator/src/index.ts compile-json",
     "changelog:compile:rss": "tsx packages/changelog-generator/src/index.ts compile-rss",
     "changelog:preview": "tsx packages/changelog-generator/src/index.ts preview",
@@ -137,9 +136,6 @@
       "last 3 firefox version",
       "last 5 safari version"
     ]
-  },
-  "engines": {
-    "node": "22.x"
   },
   "packageManager": "pnpm@10.13.1"
 }
