+++ packages/changelog-generator/package.json
@@ -11,7 +11,6 @@
     "entry:new": "tsx src/index.ts entry-new",
     "compile:json": "tsx src/index.ts compile-json",
     "compile:rss": "tsx src/index.ts compile-rss",
-    "sync:all": "tsx src/index.ts sync-all",
     "preview": "tsx src/index.ts preview",
     "publish": "tsx src/index.ts publish",
     "build:data": "pnpm run -s compile:json",
